skip waiting for prod deployments
