--- .github/workflows/pyvespa-prod-long.yml
@@ -1,29 +1,0 @@
-name: Prod integration tests - long running
-
-on:
-  workflow_dispatch:
-  schedule: # Weekly on sundays at 00:00
-    - cron: "0 0 * * 0"
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Set up Python 3.8
-        uses: actions/setup-python@v5
-        with:
-          python-version: 3.8
-
-      - name: Upgrade pip
-        run: |
-          python -m pip install --upgrade pip
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -e .[dev]
-      - name: Run tests
-        run: |
-          pytest tests/integration/test_integration_vespa_cloud_prod_long.py -s -v
