+++ tests/integration/test_integration_vespa_cloud_token.py
@@ -215,18 +215,7 @@ class TestMsmarcoProdApplicationWithTokenAuth(TestApplicationCommon):
             instance=self.instance_name,
             source_url="https://github.com/vespa-engine/pyvespa",
         )
-        # Wait for deployment to be ready
-        success = self.vespa_cloud.wait_for_prod_deployment(
-            build_no=self.build_no, max_wait=3600
-        )
-        if not success:
-            self.fail("Deployment failed")
-        self.app = self.vespa_cloud.get_application(
-            instance=self.instance_name, environment="prod"
-        )
-        self.app.wait_for_application_up(max_wait=APP_INIT_TIMEOUT)
-
-        print("Endpoint used " + self.app.url)
+        # Define fields to send and update
         self.fields_to_send = [
             {
                 "id": f"{i}",
@@ -242,6 +231,21 @@ class TestMsmarcoProdApplicationWithTokenAuth(TestApplicationCommon):
             }
             for i in range(10)
         ]
+
+    @unittest.skip(
+        "This test is too slow for normal testing. Can be used for manual testing if related code is changed."
+    )
+    def test_application_status(self):
+        # Wait for deployment to be ready
+        success = self.vespa_cloud.wait_for_prod_deployment(
+            build_no=self.build_no, max_wait=3600
+        )
+        if not success:
+            self.fail("Deployment failed")
+        self.app = self.vespa_cloud.get_application(
+            instance=self.instance_name, environment="prod"
+        )
+        self.app.wait_for_application_up(max_wait=APP_INIT_TIMEOUT)
 
     def test_using_token_endpoint(self):
         endpoint = self.app.url
@@ -275,6 +279,9 @@ class TestMsmarcoProdApplicationWithTokenAuth(TestApplicationCommon):
         self.app.delete_all_docs(
             content_cluster_name="msmarco_content", schema="msmarco"
         )
+
+    @unittest.skip("Not relevant when not waiting.")
+    def test_delete_deployment(self):
         # Deployment is deleted by deploying with an empty deployment.xml file.
         self.app_package.deployment_config = EmptyDeploymentConfiguration()
